Move authorization service to own package
